refactor(passport): isolate mutation target checks
